+++ src/project/component/types/compCollMesh.cpp
@@ -129,7 +129,7 @@ namespace Project::Component::CollMesh
       //ImGui::InputScalar("##UUID", ImGuiDataType_U64, &data.scriptUUID);
 
       int idx = modelList.size();
-      for (int i=0; i<modelList.size(); ++i) {
+      for (int i=0; i<(int)modelList.size(); ++i) {
         if (modelList[i].getUUID() == data.modelUUID.resolve(obj.propOverrides)) {
           idx = i;
           break;
@@ -139,7 +139,7 @@ namespace Project::Component::CollMesh
       auto getter = [](void*, int idx) -> const char*
       {
         auto &scriptList = ctx.project->getAssets().getTypeEntries(FileType::MODEL_3D);
-        if (idx < 0 || idx >= scriptList.size())return "<Select Model>";
+        if (idx < 0 || idx >= (int)scriptList.size())return "<Select Model>";
         return scriptList[idx].name.c_str();
       };
